+++ tsconfig.json
@@ -1,6 +1,8 @@
 {
   "compilerOptions": {
-    "outDir": "./target/js/",
+    // "outDir": "./target/js/", // cannot load ES6 modules from file: URL
+    "outFile": "./target/js/main.js",
+    "module": "system", // compatible with outFile
     "sourceMap": false,
     "strictNullChecks": true,
     "target": "es6"
